2/22
AUC: slider added, broken up summary of area and error percentage.

VUC: added section to set the axis to an arbitrary value, either along the x-axis or the y-axis.

Separate functions for the Shell Method and Disk Method to perform them in respects to the x or y axis and an axis line.